backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.22
openai==1.45.0
Pillow==12.1.1
numpy==1.26.4
SQLAlchemy==2.0.35
reportlab==4.2.2
python-dotenv==1.0.1
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
